deixando o connection dentro um arquivo separado
