+++ Semana 19/Arquitetura de Software I/src/connection/connection.ts
@@ -1,0 +1,15 @@
+import dotenv from 'dotenv'
+import knex from "knex";
+
+dotenv.config()
+
+export const connection = knex({
+    client: 'mysql',
+    connection: {
+        host: process.env.DB_HOST,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME,
+        port: 3306
+    }
+})
