+++ Semana 19/Arquitetura de Software I/src/data/deletedUser.ts
@@ -1,4 +1,4 @@
-import { connection } from "..";
+import { connection } from "../connection/connection"
 
 export default async function deletedtUser(
     id: string
