+++ Semana 19/Arquitetura de Software I/src/data/selectAllUsers.ts
@@ -1,4 +1,4 @@
-import { connection } from "..";
+import { connection } from "../connection/connection"
 
 export default async function selectAllUsers() {
     try {
