+++ Semana 19/Arquitetura de Software I/src/data/selectUserByEmail.ts
@@ -1,4 +1,4 @@
-import { connection } from ".."
+import { connection } from "../connection/connection"
 import { User } from "../types/user"
 
 export default async function selectUserByEmail(
